requirements.txt

# Core MCP and VMware dependencies
pyvmomi>=8.0.3.0.1
requests>=2.31.0
urllib3>=2.0.0

# vSphere Automation SDK
git+https://github.com/vmware/vsphere-automation-sdk-python.git

# MCP Server framework
mcp>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Ollama integration
ollama>=0.3.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Security and authentication
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# HTTP client for n8n integration
httpx>=0.25.0
